Update to latest patching method.
Owing to the need to patch BGM.XA and some other larger files, it was
easier to roll an implementation of PPF for the Lua Interface instead of
relying on armips and other batch tools.

Added:
Fix for speed-boost issue with Rune Unites.
Partial French and Italian sources (not available in patcher yet)