Fix for Emilia dialogue using hero's real name, and update Inn bug fix.
Emilia alternates between the Hero's real name and the alias.  She
should use the alias all the time for consistency.  Other oddities in
the scenario are untouched, as most characters abandon the use of
aliases early on.

The Inn Bug fix was a bit too aggressive.  Instead of removing the
premature call to RestorePartyHPMP, it needed to move to a later menu
state.  With it removed, the party would be healed by an existing,
redundant call, but only after the save data had been written.  If the
player immediately quit after staying and saving, the file would have
previous HP and MP values when loaded.