Fix issue with Rune Speed Patch
Had the directives slightly wrong in rune_speed_2.asm, and told the
assembler to start replacing one op early.  As a result real hardware or
accurate emulators could fail due to an unaligned address exception if
you used magic in the 2nd round of a battle or beyond, and the spell
chosen was anything other than the first-level spell on the Rune.  No
issues should have arisen (e.g. save corruption) from hitting this bug,
if it doesn't cause an address exception on your system.  The data the
game is attempting to store is a useless debug-trace value that Konami
left in, and it would have gone somewhere in kernel RAM.