Address issues with text fixes and clarify a description.
Changed the True Holy/Godspeed and Circuret/Circlet patches to ensure
existing strings are preserved.  One instance of "Circuret" was being
(deliberately) bypassed in an Old Book due to the search including the
terminating null.  It's now included in the patch and the remaining text
should display properly in the book.  Similarly, a Richmond
investigation includes the string "True Holy", and it was being
truncated by the insertion of a null character.

The UI description for the Tai Ho import change was altered to avoid
confusion about what's being done with Vincent in the process.