Fix crash in Gregminster events (German version)
The German translation has a path through the meeting with Lepant that traps the player into a crash.  If you choose to review the stats of the offered recruits, the option that retruns you to the actual selection dialogue is broken due to the text having too many line breaks.  The fix reformats the text, removing the line breaks.